+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add `xfail` option
+
 ## 0.5.5
 
 - Repeat `_each` and `_test` hooks when `--repeat` is specified.